feat: Err Handling:: Renamed the feature file
